+++ src/__tests__/list/manager.test.ts
@@ -352,21 +352,25 @@ describe('list', () => {
         defaultAction: '',
         loadItems: () => {
           let emitter: any = new EventEmitter()
+          let interval
+          let timeout
           emitter.dispose = () => {
             emitter.removeAllListeners()
+            clearInterval(interval)
+            clearTimeout(timeout)
           }
           if (error) {
-            setTimeout(() => {
+            timeout = setTimeout(() => {
               emitter.emit('error', new Error('error'))
               emitter.emit('end')
             }, 2)
           } else {
-            setTimeout(() => {
+            timeout = setTimeout(() => {
               emitter.emit('data', { label: 'foo' })
               emitter.emit('end')
             }, 2)
           }
-          setInterval(() => {
+          interval = setInterval(() => {
             emitter.emit('data', { label: 'bar' })
             emitter.emit('error', new Error('error'))
           }, 10)
@@ -379,7 +383,7 @@ describe('list', () => {
       error = false
       res = await manager.loadItems('emitter')
       expect(res.length).toBe(1)
-      await helper.wait(10)
+      await helper.wait(50)
     })
   })
 
@@ -410,16 +414,16 @@ describe('list', () => {
     it('should ignore <plug> insert', async () => {
       await manager.start(['location'])
       await manager.session.ui.ready
-      await nvim.eval('feedkeys("\\<plug>x", "in")')
-      await helper.wait(20)
+      await helper.listInput('<plug>')
+      await helper.listInput('x')
       expect(manager.isActivated).toBe(true)
     })
   })
 
   describe('parseArgs()', () => {
     it('should show error for bad option', async () => {
-      await helper.wait(20)
       manager.parseArgs(['$x', 'location'])
+      await helper.wait(20)
       let msg = await helper.getCmdline()
       expect(msg).toMatch('Invalid list option')
       manager.parseArgs(['-xyz', 'location'])