+++ history.md
@@ -1,3 +1,7 @@
+# 2023-01-22
+
+- `maxRestartCount` configuration for configured language server.
+
 # 2022-12-25
 
 - Create symbol tree from SymbolInformation list.